Plugins: Add MetaParser for .meta -files
This parser is intended for files that do not contain a content-part.
The parser takes the loaded file and extends it in a way, that it can
be parsed by flamingo's parse_meta_data().

I have not activated this plugin by default. It is pretty useful for my
use case. But I am quite sure that this is not an 80% use case.

Signed-off-by: Chris Fiege <chris@tinyhost.de>