sphinx_themes: version bump sphinx_rth_theme and sphinx
Until now we hold sphinx_rtd_theme at version 0.5.2, as the flamingo
integration was not compatible with newer versions.

In order to update to sphinx_rtd_theme 1.0.0, we have to add a registry
to the SphinxApp, as the HTMLThemeFactory expects to search for themes
in app.registry.html_themes now.

Update sphinx to 4.5.0, which is currently the latest and greatest.

While being there, we change the version pinning to a ">=" and try to
break & fix fast in the future.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>