sphinx_themes: set logo_url
Newer versions of the sphinx_rtd_theme (tested with 1.0.0) expect the
logo to be specified in the logo_url context variable. Provide it as a
default, in order to make it overwritable in the projects' settings.py.

Note that, in contrast to Sphinx, flamingo doesn't automatically set the
'logo' context variable, so you must provide SPHINX_THEMES_LOGO=True in
order to show the logo at all.

Example, from a project's settings.py:

  # workaround: if 'logo' isn't defined, the sphinx_rtd_theme doesn't
  # render the logo at all; if we set it to something defined, 'logo_url'
  # below is actually used
  SPHINX_THEMES_LOGO = True
  SPHINX_THEMES_LOGO_URL = '/static/mylogo.svg'

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>