tox: make it possible to run a specified test
Currently pytest.ini hardcodes the tests to 'tests', which overwrites
anything specified on the pytest commandline, and in result, it is not
possible to run

  tox -- <testname>

to run a single test.

Remove 'tests' from pytest.ini and add {posargs} in tox.ini, so any
additional commandline arguments are propagated to pytest and not
ignored there.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>