packaging: Move release-workflow from tox to make
This changes the way pypi releases of this packages are made.
Until now sdists and uploads have been managed by tox.
With this change creating the sdist and uploading are managed by make
using the Makefile.
With this change we also split the sdist- and upload-step. This allows
some room for reviews of the created sdist.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>