tox: move setenv=FLAMINGO_TEST=1 to "testenv:.pkg" section
Since tox 4, "packaging environments no longer inherit their settings
from the testenv section" [1]. This becomes an issue for flamingo
because the environment variable FLAMINGO_TEST is checked in its
setup.py.

Fix this by moving the setenv directive to the build environment
section. passenv= in flamingo is not affected because other env lookups
are performed outside of its setup.py.

[1] https://tox.wiki/en/4.11.3/upgrading.html#packaging-configuration-and-inheritance

Signed-off-by: Bastian Krause <bst@pengutronix.de>