Context: Filter MarkupResemblesLocatorWarning from BeautifulSoup
BeautifulSoup will raise a `MarkupResemblesLocatorWarning` every time it
is asked to parse a string that looks like a path or url.

This warning is intended to aid novice users. But in the context of
flamingo it leads to warning every time a content consists of something
that looks like a path or url to bs4.

With this change we globally disable this specific warning from bs4.
Disabling warning this early in the startup process makes this effective
for flamingo itself and all plugins. (Settings are loaded earlier - but
I hope nobody needs to run bs4 in their settings modules...)

Signed-off-by: Chris Fiege <cfi@pengutronix.de>