plugins/feeds: render summary and make URLs in it absolute
Unrendered summaries contain directives such as:

  {{ link('content/blog-post.html', 'foo') }}

These do not belong in the final HTML output. A function for summary
rendering can be provided by a custom summary plugin.
Use that if available and make rendered URLs absolute afterwards.

See [1] for an example summary plugin. But note that it contains
theme-related information.

[1] https://github.com/pengutronix/flamingo-ptx-blog-engine/blob/master/flamingo_ptx_blog_engine/summary.py

Signed-off-by: Bastian Krause <bst@pengutronix.de>