plugins/redirects: Fix tabs are parsed as part of the source
Until now the columns of a `.rr`-files could only be separated by
spaces.
This regularly leads to not working redirects, where the generated
source-file would contain tabs as part of the file name.

This comes down to how the regex parses the file: Before this change
only spaces were not part of source and destination - but other
whitespace was ok.

With this change source and destination can not contain any whitespace.
Thus tabs are parsed as separators - as one would expect.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>